requirements.txt
# 核心依赖
pywin32>=306; sys_platform == 'win32'
psutil>=5.9.0
Pillow>=10.0.0

# OCR相关
requests>=2.28.0

# 截图相关
windows-capture>=0.4.0; sys_platform == 'win32'

# 配置管理
python-dotenv>=1.0.0

# 测试框架
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
